On Load, check for cores required by splice and report if any are missing, refusing to load.
Better than an ugly error. Added preference to disable this check in case it causes problems. Unlikely since the logic is directly taken from the canApplyAffine() check that threw the aforementioned ugly error, so if one fails the other should too. Also resolved mind-bending issue in Prefs resolved by added contain() method. Was incorrectly using Prefs.get() to see if key was in prefs, but get() returns the *value* which is a boolean in this case, causing the default value to be re-set every time. I believe this is a record-long commit message!